Refactor about where the objects should be in of model module.
