+++ pymock_api/server/sgi/__init__.py
@@ -3,22 +3,5 @@
 The factory to generate SGI application instance to run Python web framework.
 """
 
-from argparse import Namespace
-from typing import Optional
-
-from ._model import Command, CommandOptions, Deserialize, ParserArguments
 from .cmd import ASGIServer, WSGIServer
 from .cmdoption import ASGICmdOption, WSGICmdOption
-
-
-def deserialize_parser_args(args: Namespace, subcmd: Optional[str] = None) -> ParserArguments:
-    """Deserialize the object *argparse.Namespace* to *ParserArguments*.
-
-    Args:
-        args (Namespace): The arguments which be parsed from current command line.
-
-    Returns:
-        A *ParserArguments* type object.
-
-    """
-    return Deserialize.parser_arguments(args, subcmd)
