+++ test/unit_test/server/sgi/model.py
@@ -1,24 +1,14 @@
-from argparse import Namespace
-from typing import Type
 from unittest.mock import Mock, patch
 
 import pytest
 
-from pymock_api.server.sgi._model import (
-    Command,
-    CommandOptions,
-    Deserialize,
-    ParserArguments,
-)
+from pymock_api.server.sgi._model import Command, CommandOptions
 
 from ...._values import (
     _Bind_Host_And_Port,
     _Cmd_Option,
     _Log_Level,
-    _Test_App_Type,
-    _Test_Config,
     _Test_Entry_Point,
-    _Test_SubCommand,
     _Workers_Amount,
 )
 
@@ -81,28 +71,3 @@ class TestCommand:
     def expected_cmd_line(cls, command: Command) -> str:
         host_and_port, workers, log_level = _get_cmd_options()
         return " ".join([command.entry_point, host_and_port, workers, log_level, command.app_path])
-
-
-class TestDeserialize:
-    @pytest.fixture(scope="function")
-    def deserialize(self) -> Type[Deserialize]:
-        return Deserialize
-
-    def test_parser_arguments(self, deserialize: Type[Deserialize]):
-        namespace_args = {
-            _Test_SubCommand: _Test_SubCommand,
-            "config": _Test_Config,
-            "app_type": _Test_App_Type,
-            "bind": _Bind_Host_And_Port.value,
-            "workers": _Workers_Amount.value,
-            "log_level": _Log_Level.value,
-        }
-        namespace = Namespace(**namespace_args)
-        arguments = deserialize.parser_arguments(namespace, subcmd=_Test_SubCommand)
-        assert isinstance(arguments, ParserArguments)
-        assert arguments.subparser_name == _Test_SubCommand
-        assert arguments.config == _Test_Config
-        assert arguments.app_type == _Test_App_Type
-        assert arguments.bind == _Bind_Host_And_Port.value
-        assert arguments.workers == _Workers_Amount.value
-        assert arguments.log_level == _Log_Level.value
